lang/openjdk21: Workaround for the VM dying on Apple Silicon chips
See the patch for threadWXSetters.inline.hpp for details. I rebuilt the
bootkit for aarch64 with new patches applied. This may cause some
performance regression on Cortex series, but as I stated in the patch
comment, relying on implementation details of a particular chip is
fundamentally unsound, and that's why it didn't work on Apple Silicon in
the first place.

But since the upstream has stated they are knowingly doing this, the chance
of these patches getting accepted is slim, if not zero.