print/podofo: Update HOMEPAGE
podofo has moved to github.  Adjust HOMEPAGE and leave note that it
needs updating to 1.0.1, which needs c++17.

podofo 0.9.8 doesn't build with cmake 4, but as it needs an update
anyway which might avoid that (because upstream might have fixed the
bug or performed a cmake exorcism), that isn't fixed in this commit.