py-pspdfutils: update to 3.3.12.
This release requires Python 3.12 or newer.

All of the commands are fixed to allow the use of a hyphen - to
mean standard input (for an input argument) or standard output (for
an output argument), as documented. Thanks very much to Olivier
Scherler for the bug report and fix for all but psjoin.

The obsolete old scripts, which were not installed as part of the
package but kept in “old-scripts” for backwards compatibility, are
removed: they were deprecated a long time ago.

The installation instructions are updated to recommend the use of
pipx rather than pip.

The Makefile has been updated.