py-textual: updated to 6.0.0
6.0.0

This is a fairly large update, with some new features and optimizations.

There are a few breaking changes, which are unlikely to impact many apps. Unless you have build custom line-API widgets. See below for the details.

You may have to regenerate your snapshot tests, mostly as a result of the optimizations. I wouldn't expect the changes result in any material changes.