mail/mailutils: update to 3.20
# noteworthy changes (since 3.19)

* movemail

** Synchronization mode.

Setting synchronization mode allows the user to keep messages in remote
source mailbox, while downloading only recently received messages.
The mode is defined via the "--sync" command line option or "sync"
configuration statement.  Allowed values are:

  uidnext
     Use the combination of uidvalidity/uidnext values.  This is useful
     mainly if the source mailbox is accessed via IMAP4 protocol.  When
     using this method, movemail stores session metadata in files in the
     directory ~/.movemail.sync.  The directory location can be changed
     using the "--sync-dir" option or "sync-dir" configuration statement.

  uidl
     Use UIDL values.  Useful if the source mailbox is accessed via the
     POP3 protocol.

  all
     Download each message.  Provided for completness sake.  This is
     the default behavior when no --sync option was supplied.

** The --reverse option removed

The option made little sense for movemail and was never used.

** The --max-messages option


This option sets the maximum number of latest messages to process.

* New Sieve test: uidnew

The "uidnew" test keeps track of the processed messages.  It evaluates
to true if the current message was not processed before.

The test is implemented as an external module.  To require it, use

  require "test-uidnew";

Sample use

  if uidnew
    {
      fileinto "store"; keep;
    }

For each processed mailbox, the test keeps its state in a GDBM file
~/.uidnew.db.  The ":db" tagged argument can be used to alter this
location.

* Bugfixes

** imap4d UID SEARCH command

The UID SEARCH command incorrectly treated message range argument as
UIDs instead of as message sequence numbers.