k6: Update to 1.2.3
Changes:
1.2.3
k6 1.2.3 is a small patch with a couple of bug fixes


1.2.2
k6 1.2.2 is a small patch release fixing a panic and two other smaller bugfixes.


1.2.0 and 1.2.1
k6 v1.2.1 is here! This release includes:

- Automatic extension resolution (previously Binary Provisioning)
  enabled for everyone
- gRPC gets better handling of `NaN` and `Infinity` float values and
  easier health check
- Browser module gets `page.route`, all the `page.getBy*` APIs,
  `locator.all()`, and `page.waitForURL`


1.1.0
k6 `v1.1.0` is here! This release includes:

- New `count`, `nth`, `first`, and `last` methods for the browser
  module's Locator API
- The `k6/experimental/webcrypto` module has been removed as its
  functionality is available globally.
- Group results in the `full` end-of-test summary are now sorted as in
  code and properly indented.


1.0
Grafana k6 v1.0 is here!

After 9 years of iteration and countless community contributions, we're
thrilled to announce Grafana k6 v1.0.

While many features and capabilities in this release were introduced
gradually in previous versions, k6 v1.0 marks a turning point: a
commitment to stability, formal support guarantees, and transparency in
how we evolve and develop the project from here. This milestone is more
than a version number; it's about trust, reliability, and empowering
you to test confidently.

What's new?
1. Stability You Can Build On
2. First-Class TypeScript Support
3. Extensions Made Simple
4. Revamped test summary
5. Quality of Life Upgrades


0.57.0
k6 v0.57.0` is here! This release includes:

- Introducing helpers for functional testing.
- The `k6 new` command now supports templates and ProjectIDs.
- The `k6/experimental/csv` module gets a new `asObjects` option.
- We no longer support the `k6/experimental/browser` module, in favor
  of `k6/browser`.
- Moving most of non-public APIs to the `internal` package.