nodejs22: updated to 22.20.0
22.20.0 'Jod' (LTS)

Notable Changes

OpenSSL updated to 3.5.2

For official Node.js builds, or builds using the default build configuration, Node.js now bundles OpenSSL 3.5.2. This update allows Node.js 22.x to be supported through to the planned End-of-Life date of 2027-04-30 as the previously bundled OpenSSL 3.0.x goes out of support in September 2026.

This change does not affect third-party builds of Node.js that link to an external OpenSSL (or OpenSSL-compatible) library.

Other notable changes

- crypto: update root certificates to NSS 3.114 (Node.js GitHub Bot)
- doc: stabilize --disable-sigusr1 (Rafael Gonzaga)
- doc: mark path.matchesGlob as stable (Aviv Keller)
- (SEMVER-MINOR) http: add Agent.agentKeepAliveTimeoutBuffer option (Haram Jeong)
- (SEMVER-MINOR) http2: add support for raw header arrays in h2Stream.respond() (Tim Perry)
- inspector: add http2 tracking support (Darshan Sen)
- (SEMVER-MINOR) sea: implement execArgvExtension (Joyee Cheung)
- (SEMVER-MINOR) sea: support execArgv in sea config (Joyee Cheung)
- (SEMVER-MINOR) stream: add brotli support to CompressionStream and DecompressionStream (Matthew Aitken)
- (SEMVER-MINOR) test_runner: support object property mocking (Idan Goshen)
- (SEMVER-MINOR) worker: add cpu profile APIs for worker (theanarkh)