tealdeer: update to 1.8.0
v1.8.0 (2025-10-03)

One year and one day have passed since tealdeer version 1.7.0 was
released, so it's time for an update! Tealdeer 1.8 comes with a complete
rewrite of the page cache and contains many long awaited improvements
around it.

Firstly, tealdeer now supports language-specific downloads. This means
that only the pages matching the configured languages are downloaded
when updating the cache. The languages used for searching pages can be
configured separately to the ones used for updating, so it is possible
to download pages in languages that are not usually queried.

Next to configuring which languages are used for searching, it is now
also possible to specify which platforms are used in the config file.
Importantly, the default behavior for page search has changed so that
all platforms are searched if no page is found for the platform that
tealdeer is running on. To restore the behavior of tealdeer 1.7, users
should set

[search]
platforms = ["current", "common"]

in their config file.

Coming back to updating, the default build configuration of tealdeer now
includes multiple TLS backends. This means that tealdeer does not have
to be rebuilt to try out a different TLS backend. The used backend can
be chosen in the config file. By default, tealdeer comes with support
for rustls using webpki certificates or system certificates. Native TLS
is supported, but not enabled by default to avoid build troubles with
OpenSSL and musl.