fix build with GCC 14.
GCC 14 revealed several problems with transocde.

the ImageMagick7 port was never complete, and there were missing symbols
that would make the plugins fail to load.  disable these components for
now, until someone spends the time to complete the port.

disable the resample filter which uses APIs removed in ffmpeg4.

provided copies of 2 deprecated lame function prototypes that are always
hidden in lame.h, but are compiled into the library.