fix build with GCC 14.
this one is annoying.  initially, it fails as it force sets "CFLAGS=-O"
i an sub-make that tries to compile invalid in gcc 14 default code.  the
patch to src/Makefile is to pass down CFLAGS as CFLAGS.

the "configure"-alike setup here absolutely requires to turn off the new
errors as it explicitly abuses that old C feature to find what symbols
do or don't exist.

as such, just add -Wno-implicit-int -Wno-implicit-function-declaration
to CFLAGS with GCC 14, and this builds again.